Clock done
Added JavaScript to find the time, and a function to calculate the time from the current time to Jan 1st. The function rounds down the clock so that each time unit appears as a whole number, but it is still accurate. The function finds the time between the two dates, and multiplies it by a number to put it in the correct format. The function occurs after a set interval, so that it updates each second.